// models/detalhesEvento.ts
interface ChapterOrganizer {
    title: string;
    role: string;
    order: number;
    first_name: string;
    last_name: string;
    twitter: string;
    company: string;
    bio: string;
    avatar: string;
    id: number;
    profile_url: string;
}

export interface DetalhesEvento {
    agenda: string;
    allow_new_agenda: boolean;
    allow_registration_as_a_guest: boolean;
    audience_type: string;
    banner: string | null;
    chapter_banner: string;
    chapter_id: number;
    chapter_slug: string;
    chapter_team_list: ChapterOrganizer[];
    chapter_title: string;
    chapter_url: string;
    cohost_registration_url: string;
    completed: boolean;
    cropped_banner_url: string;
    cropped_picture_url: string;
    currency: string;
    custom_event_url: string | null;
    custom_tickets_url: string;
    description: string;
    description_short: string;
    end_date: string;
    end_date_iso: string;
    event_banner: string;
    event_timezone: string;
    event_type_title: string;
    featured_attendees: any[];
    hide_agenda_on_event_page: boolean;
    id: number;
    is_hidden: boolean;
    is_virtual_event: boolean;
    join_virtual_event_url: string;
    minutes_until_start: number;
    payment_methods: any[];
    picture: string;
    registration_required: boolean;
    rsvp_only: boolean;
    sharing_disabled: boolean;
    show_map: boolean;
    show_short_description_on_event_banner: boolean;
    slideshare_url: string | null;
    slug: string;
    start_date: string;
    start_date_iso: string;
    static_url: string;
    tags: string[];
    timezone_abbreviation: string;
    title: string;
    total_attendees: number;
    url: string;
    use_external_ticketing: boolean;
    use_featured_attendees: boolean;
    venue_address: string | null;
    venue_city: string | null;
    venue_name: string | null;
    venue_zip_code: string | null;
    video_url: string | null;
    virtual_event_instructions: string;
    virtual_event_minutes_before_can_join: number;
    virtual_event_ready_date: string;
    virtual_venue_link: string;
}